Make each waiter thread a singleton.
Ensure that only one instance of each type can run at a time.  If one
trying to be created has a different instance_id than the one currently
running, throw an exception.  It would be nice to be able to cancel the
currently running thread and run the new one instead, but Python doesn't
seem to support thread cancellation (except through nasty hacks).