Add locking to SingletonThread.
Due the the vagaries of python's class attributes, if a lock is defined
as a class attribute in the base class, it will be the same lock
instance for all derived classes.  I had hoped to lazy-create a lock
per-derived-type using __dict__.setdefault, but unfortunately class
objects don't use real dicts, and don't have a setdefault method on what
they do use.  So, use a 'meta' lock defined on the base class to
protect the lazy creation of per-derived-type locks.