Fix empty File menu on Mac.
use platform="win|unix" to hide file menu on Mac, but add an additional
Exit menu item to the Edit menu with platform="mac", cause apparently
Mac needs the item to exist SOMEWHERE even though it's not going to show
it there.  Grr.

Use stock IDs for more items that turn out to actually have stock IDs,
and don't specify properties of these items that are/can be inherited
from the stock items.  This also gets rid of the refresh icon.