Tweak userdata powershell script.
After turning off the download progress bar, the Epic Games Launcher
install was failing, presumably due to timing/conflicting with the NICE
dcv install which was still running.  Tweaked to download both Epic
Games and Steam while NICE is installing, and then wait for NICE to
finish before moving on and installing Epic.